fix active session test by adding a dedicated NotificationManager This NotificationManager is dependency injected by dagger hilt which at the same time creates the notification channels. Thereby the whole mechanism works out of the box when used in tests.
