+++ app/src/androidTest/java/app/musikus/activesession/presentation/ActiveSessionScreenTest.kt
@@ -10,6 +10,7 @@ package app.musikus.activesession.presentation
 
 import androidx.activity.compose.setContent
 import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.assertIsNotDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
@@ -23,11 +24,13 @@ import app.musikus.library.data.entities.LibraryItemCreationAttributes
 import app.musikus.library.domain.usecase.LibraryUseCases
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import javax.inject.Inject
+import kotlin.time.Duration.Companion.seconds
 
 @HiltAndroidTest
 class ActiveSessionScreenTest {
@@ -77,5 +80,34 @@ class ActiveSessionScreenTest {
 
         composeRule.onNodeWithText("TestItem1").performClick()
         composeRule.onNodeWithContentDescription("Next item").assertIsDisplayed()
+    }
+
+    @Test
+    fun pauseAndResumeSession() {
+        // Start session
+        composeRule.onNodeWithContentDescription("Start practicing").performClick()
+        composeRule.onNodeWithText("TestItem1").performClick()
+
+        // Pause session after advancing time by 1 second
+        // (a session with 0 seconds can not be paused)
+        fakeTimeProvider.advanceTimeBy(1.seconds)
+        composeRule.onNodeWithContentDescription("Pause").performClick()
+
+        // Pause timer is displayed
+        composeRule.onNodeWithText("Paused 00:00").assertIsDisplayed()
+
+        fakeTimeProvider.advanceTimeBy(90.seconds)
+
+        runBlocking {
+            delay(100)
+        }
+
+        // Pause timer shows correct time
+        composeRule.onNodeWithText("Paused 01:30")
+            .assertIsDisplayed()
+            .performClick() // Resume session
+
+        // Pause timer is hidden
+        composeRule.onNodeWithText("Paused", substring = true).assertIsNotDisplayed()
     }
 }
