+++ app/src/main/java/app/musikus/activesession/presentation/ActiveSessionScreen.kt
@@ -808,7 +808,10 @@ private fun ActiveSessionTopBar(
                     IconButton(
                         onClick = onDiscard,
                     ) {
-                        Icon(imageVector = Icons.Outlined.Delete, contentDescription = null)
+                        Icon(
+                            imageVector = Icons.Outlined.Delete,
+                            contentDescription = stringResource(id = R.string.active_session_top_bar_delete)
+                        )
                     }
 
                     TextButton(
@@ -831,7 +834,7 @@ private fun PauseButton(
     ) {
         Icon(
             imageVector = Icons.Filled.Pause,
-            contentDescription = null
+            contentDescription = stringResource(id = R.string.active_session_top_bar_pause)
         )
     }
 }
@@ -987,7 +990,12 @@ private fun PracticeTimer(
                         contentColor = MaterialTheme.colorScheme.onTertiary
                     )
                 ) {
-                    Icon(imageVector = Icons.Outlined.PlayCircle, contentDescription = null)
+                    Icon(
+                        imageVector = Icons.Outlined.PlayCircle,
+                        contentDescription = stringResource(
+                            id = R.string.active_session_timer_subheading_resume
+                        )
+                    )
                     Spacer(Modifier.width(MaterialTheme.spacing.small))
                     Text(text = uiState.value.subHeadingText.asString())
                 }
