+++ app/src/main/java/app/musikus/activesession/presentation/SessionService.kt
@@ -71,6 +71,9 @@ class SessionService : Service() {
     lateinit var useCases: ActiveSessionUseCases
     private var _timer: Timer? = null
 
+    @Inject
+    lateinit var notificationManager: NotificationManager
+
     /** Broadcast receiver (currently only for pause action) */
     private val myReceiver: BroadcastReceiver = object : BroadcastReceiver() {
         override fun onReceive(context: Context, intent: Intent) {
@@ -186,8 +189,7 @@ class SessionService : Service() {
     private fun updateNotification() {
         Log.d(LOG_TAG, "updateNotification")
         applicationScope.launch {
-            val mNotificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-            mNotificationManager.notify(SESSION_NOTIFICATION_ID, createNotification())
+            notificationManager.notify(SESSION_NOTIFICATION_ID, createNotification())
         }
     }
