+++ app/src/main/java/app/musikus/core/di/NotificationModule.kt
@@ -1,0 +1,33 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * Copyright (c) 2024 Matthias Emde
+ */
+
+package app.musikus.core.di
+
+import android.app.NotificationManager
+import android.content.Context
+import app.musikus.core.presentation.createNotificationChannels
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NotificationModule {
+
+    @Provides
+    @Singleton
+    fun provideNotificationManager(
+        @ApplicationContext context: Context
+    ): NotificationManager {
+        createNotificationChannels(context)
+        return context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+    }
+}
